backend/requirements.txt
# --- Core API ---
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy
psycopg2-binary
python-multipart>=0.0.9
python-dotenv
httpx
requests

# --- Security & Auth ---
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
authlib>=1.3.0
itsdangerous==2.2.0
email-validator

# --- AI & External Services ---
google-genai
resend

# --- File & Document Handling ---
fpdf2
pillow
python-bidi
arabic-reshaper
boto3>=1.34.0

# --- Templates ---
jinja2

# --- WebSockets ---
websockets

# --- Redis Queue (PRO ARCHITECTURE) ---
redis>=5.0.0
rq>=1.15.0
slowapi==0.1.9